chore(core): move all abstractions to core
