+++ Cargo.toml
@@ -15,9 +15,9 @@ channel = { path = "crates/channel", package = "walrus-channel", version = "0.0.
 cli = { path = "crates/cli", package = "openwalrus", version = "0.0.2" }
 daemon = { path = "crates/daemon", package = "walrus-daemon", version = "0.0.2" }
 mcp = { path = "hook/mcp", package = "walrus-mcp", version = "0.0.2" }
-memory = { path = "hook/memory", package = "walrus-memory", version = "0.0.2" }
+memory = { path = "crates/memory", package = "walrus-memory", version = "0.0.2" }
 model = { path = "crates/model", package = "walrus-model", version = "0.0.2" }
-protocol = { path = "crates/protocol", package = "walrus-protocol", version = "0.0.2" }
+socket = { path = "crates/socket", package = "walrus-socket", version = "0.0.2" }
 skill = { path = "hook/skill", package = "walrus-skill", version = "0.0.2" }
 wcore = { path = "crates/core", package = "walrus-core", version = "0.0.2" }
 wcron = { path = "hook/cron", package = "walrus-cron", version = "0.0.2" }
@@ -74,7 +74,7 @@ strip = true
 
 [workspace.metadata.conta]
 packages = [
-    "protocol",
+    "socket",
     "wcore",
     "channel",
     "wcron",
