+++ crates/cli/Cargo.toml
@@ -23,7 +23,8 @@ cuda = ["daemon/cuda"]
 
 [dependencies]
 daemon = { workspace = true, optional = true }
-protocol = { workspace = true, features = ["socket"] }
+socket = { workspace = true, features = ["socket"] }
+wcore.workspace = true
 
 # crates.io
 anyhow.workspace = true
