+++ crates/core/src/memory/embedder.rs
