+++ crates/core/src/protocol/api/client.rs
@@ -1,6 +1,6 @@
 //! Client trait — transport primitives plus typed provided methods.
 
-use crate::message::{
+use crate::protocol::message::{
     AgentDetail, AgentInfoRequest, AgentList, ClearSessionRequest, DownloadEvent, DownloadRequest,
     GetMemoryRequest, McpAddRequest, McpAdded, McpReloaded, McpRemoveRequest, McpRemoved,
     McpServerList, MemoryEntry, MemoryList, SendRequest, SendResponse, SessionCleared,
