+++ crates/daemon/Cargo.toml
@@ -21,7 +21,7 @@ channel.workspace = true
 mcp.workspace = true
 memory.workspace = true
 model.workspace = true
-protocol = { workspace = true, features = ["socket"] }
+socket = { workspace = true, features = ["socket"] }
 skill.workspace = true
 wcore.workspace = true
 wcron.workspace = true
