+++ crates/daemon/src/daemon/event.rs
@@ -8,11 +8,11 @@
 use crate::daemon::Daemon;
 use compact_str::CompactString;
 use futures_util::{StreamExt, pin_mut};
-use protocol::{
+use tokio::sync::{mpsc, oneshot};
+use wcore::protocol::{
     api::Server,
     message::{client::ClientMessage, server::ServerMessage},
 };
-use tokio::sync::{mpsc, oneshot};
 use wcron::CronJob;
 
 /// Inbound event from any source, processed by the central event loop.
