+++ crates/memory/sql/delete.sql
