+++ crates/memory/sql/recall_fts.sql
