+++ crates/memory/sql/recall_vector.sql
