+++ crates/memory/sql/select_entries.sql
