+++ crates/memory/sql/select_entry.sql
