+++ crates/memory/sql/select_value.sql
