+++ crates/memory/sql/upsert_full.sql
