+++ crates/memory/src/utils.rs
