+++ crates/socket/README.md
