+++ crates/socket/src/socket/codec.rs
