+++ hook/cron/Cargo.toml
@@ -11,7 +11,6 @@ repository.workspace = true
 keywords.workspace = true
 
 [dependencies]
-protocol.workspace = true
 wcore.workspace = true
 
 # crates.io
