+++ hook/cron/src/lib.rs
@@ -7,15 +7,15 @@
 use chrono::Utc;
 use compact_str::CompactString;
 use cron::Schedule;
-use protocol::{
-    api::{Client, Server},
-    message::SendRequest,
-};
 use std::{str::FromStr, sync::Arc};
 use tokio::{
     sync::{RwLock, broadcast, mpsc},
     task::JoinHandle,
     time,
+};
+use wcore::protocol::{
+    api::{Client, Server},
+    message::SendRequest,
 };
 
 mod client;
